Cleanup image load code.
