+++ cmd/cri-containerd/cri_containerd.go
@@ -24,22 +24,16 @@ import (
 	"net/http/pprof"
 	"os"
 	"os/signal"
-	"path/filepath"
 	"runtime"
-	"strings"
 	"syscall"
 
 	"github.com/docker/docker/pkg/reexec"
 	"github.com/golang/glog"
 	"github.com/opencontainers/selinux/go-selinux"
-	"github.com/renstrom/dedent"
 	"github.com/spf13/cobra"
-	"golang.org/x/net/context"
 	"k8s.io/kubernetes/pkg/util/interrupt"
 
 	"github.com/kubernetes-incubator/cri-containerd/cmd/cri-containerd/options"
-	api "github.com/kubernetes-incubator/cri-containerd/pkg/api/v1"
-	"github.com/kubernetes-incubator/cri-containerd/pkg/client"
 	"github.com/kubernetes-incubator/cri-containerd/pkg/server"
 	"github.com/kubernetes-incubator/cri-containerd/pkg/version"
 )
@@ -59,30 +53,6 @@ var cmd = &cobra.Command{
 	Short: "A containerd based Kubernetes CRI implementation.",
 	Long:  desc,
 }
-
-var (
-	loadLong = dedent.Dedent(`
-		Help for "load TAR" command
-
-		TAR - The path to a tar archive containing a container image.
-
-		Requirement:
-			Containerd and cri-containerd daemons (grpc servers) must already be up and
-			running before the load command is used.
-
-		Description:
-			Running as a client, cri-containerd implements the load command by sending the
-			load request to the already running cri-containerd daemon/server, which in
-			turn loads the image into containerd's image storage via the already running
-			containerd daemon.`)
-
-	loadExample = dedent.Dedent(`
-		- use docker to pull the latest busybox image and save it as a tar archive:
-			$ docker pull busybox:latest
-			$ docker save busybox:latest -o busybox.tar
-		- use cri-containerd to load the image into containerd's image storage:
-			$ cri-containerd load busybox.tar`)
-)
 
 // Add golang flags as persistent flags.
 func init() {
@@ -107,38 +77,6 @@ func versionCommand() *cobra.Command {
 			version.PrintVersion()
 		},
 	}
-}
-
-func loadImageCommand() *cobra.Command {
-	c := &cobra.Command{
-		Use:     "load TAR",
-		Long:    loadLong,
-		Short:   "Load an image from a tar archive.",
-		Args:    cobra.ExactArgs(1),
-		Example: loadExample,
-	}
-	c.SetUsageTemplate(strings.Replace(c.UsageTemplate(), "Examples:", "Example:", 1))
-	endpoint, timeout := options.AddGRPCFlags(c.Flags())
-	c.RunE = func(cmd *cobra.Command, args []string) error {
-		cl, err := client.NewCRIContainerdClient(*endpoint, *timeout)
-		if err != nil {
-			return fmt.Errorf("failed to create grpc client: %v", err)
-		}
-		path, err := filepath.Abs(args[0])
-		if err != nil {
-			return fmt.Errorf("failed to get absolute path: %v", err)
-		}
-		res, err := cl.LoadImage(context.Background(), &api.LoadImageRequest{FilePath: path})
-		if err != nil {
-			return fmt.Errorf("failed to load image: %v", err)
-		}
-		images := res.GetImages()
-		for _, image := range images {
-			fmt.Println("Loaded image:", image)
-		}
-		return nil
-	}
-	return c
 }
 
 func main() {
